+++ sword-tests/src/lib.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod request {
     mod cookies;
+    mod extractors;
     mod multipart;
     mod query;
 }